Initialize OreLimiter/AntiSM before loading worlds
This fixes a long standing bug rarely breaking AntiStripmine. When a world is pregenerated not within the same server cycle as the initial creation, AntiStripmine would not receive the WorldInitEvent (as the worlds would be loaded before the listener being registered), so the world would never be registered in the AntiStripmine class.

OreLimiter is technically not affected by this (using ChunkPopulateEvent), but better to be safe than sorry.